Small UI update Small change to service search behaviour and interaction Fixed bug that cause replica services to appear Added network activity indicator

+++ FlyByWire/ABNetworkController.m
@@ -49,7 +49,7 @@
 - (void)stream:(NSStream *)aStream handleEvent:(NSStreamEvent)eventCode
 {
     uint8_t buffer[1024];
-    int len;
+    NSInteger len;
 	
     switch (eventCode) {
